pull request

+++ build/build_release.sh
@@ -240,28 +240,6 @@ else
   gwc_tag=`cat src/pom.xml | grep "<gwc.version>" | sed 's/ *<gwc.version>\(.*\)<\/gwc.version>/\1/g'`
 fi
 
-# update the release notes
-notes=RELEASE_NOTES.txt
-pushd src/release > /dev/null
-sed -i "s/@VER@/$tag/g" $notes
-sed -i "s/@DATE@/`date "+%b %d, %Y"`/g" $notes
-sed -i "s/@JIRA_VER@/$jira_id/g" $notes
-
-gt_ver_info=$gt_tag
-if [ ! -z $gt_rev ]; then
-  gt_ver_info="$gt_ver_info, rev $gt_rev"
-fi
-
-gwc_ver_info=$gwc_tag
-if [ ! -z $gwc_rev ]; then
-  gwc_ver_info="$gwc_ver_info, rev $gwc_rev"
-fi
-
-sed -i "s/@GT_VER@/$gt_ver_info/g" $notes
-sed -i "s/@GWC_VER@/$gwc_ver_info/g" $notes
-
-popd > /dev/null
-
 # update version numbers
 old_ver=`get_pom_version src/pom.xml`
 
@@ -285,22 +263,49 @@ if [ -z $SKIP_BUILD ]; then
   # build the javadocs
   mvn javadoc:aggregate
 
-  # build the user docs
-  pushd ../doc/en/user > /dev/null
-  make clean html
-  make latex
-  cd build/latex
-  sed  "s/includegraphics/includegraphics[scale=0.5]/g" GeoServerUserManual.tex > manual.tex
-  # run pdflatex twice in a row to get the TOC, and ignore errors 
-  set +e
-  pdflatex -interaction batchmode manual.tex
-  pdflatex -interaction batchmode manual.tex
-  set -e
+  ##################
+  # Build the docs
+  ##################
 
-  cd ../../../developer
-  make clean html
+
+
+  pushd ../doc/en > /dev/null
+
+  # 2.11 and older uses make
+  if [ -e user/Makefile ]
+  then
+    # build the user docs
+    cd user
+    make clean html
+    make latex
+    cd build/latex
+
+    sed  "s/includegraphics/includegraphics[scale=0.5]/g" GeoServerUserManual.tex > manual.tex
+    # run pdflatex twice in a row to get the TOC, and ignore errors 
+    set +e
+    pdflatex -interaction batchmode manual.tex
+    pdflatex -interaction batchmode manual.tex
+    set -e
+
+    if [ ! -f manual.pdf ]; then
+      echo "Failed to build pdf manual. Printing latex log:"
+      cat manual.log
+    fi
+
+    # build the developer docs
+    cd ../../../developer
+    make clean html
+
+  # 2.12 and newer uses ant to do everything
+  else
+    ant clean user -Dproject.version=$tag
+    ant user-pdf -Dproject.version=$tag
+    ant developer -Dproject.version=$tag
+  fi
 
   popd > /dev/null
+
+
 fi
 
 mvn $MAVEN_FLAGS assembly:attached
@@ -335,8 +340,17 @@ if [ -e developer ]; then
   unlink developer
 fi
 
-ln -sf ../../../doc/en/user/build/html user
-ln -sf ../../../doc/en/developer/build/html developer
+# paths for 2.12 and newer docbuild
+usertarget=target/user
+devtarget=target/developer
+# paths for 2.11 and older docbuild
+if [ -e ../../../doc/en/user/Makefile ]; then
+  usertarget=user/build
+  devtarget=developer/build
+fi
+
+ln -sf ../../../doc/en/$usertarget/html user
+ln -sf ../../../doc/en/$devtarget/html developer
 ln -sf ../../../doc/en/release/README.txt readme
 
 htmldoc=geoserver-$tag-htmldoc.zip
@@ -351,18 +365,12 @@ unlink readme
 popd > /dev/null
 
 echo "copying artifacts to $dist"
-cp $artifacts/../../../doc/en/user/build/latex/manual.pdf $dist/geoserver-$tag-user-manual.pdf
 cp $artifacts/*-plugin.zip $dist/plugins
 for a in `ls $artifacts/*.zip | grep -v plugin`; do
   cp $a $dist
 done
 
-# fire off mac and windows build machines
-if [ -z $SKIP_INSTALLERS ]; then
-  echo "starting installer jobs"
-  start_installer_job $WIN_JENKINS $WIN_JENKINS_USER $WIN_JENKINS_KEY $tag
-  start_installer_job $MAC_JENKINS $MAC_JENKINS_USER $MAC_JENKINS_KEY $tag
-fi
+cp $artifacts/../../../doc/en/$usertarget/latex/manual.pdf $dist/geoserver-$tag-user-manual.pdf
 
 # git commit changes on the release branch
 pushd .. > /dev/null
@@ -372,6 +380,13 @@ init_git $git_user $git_email
 git add . 
 git commit -m "updating version numbers and release notes for $tag" .
 popd > /dev/null
+
+# fire off mac and windows build machines
+if [ -z $SKIP_INSTALLERS ]; then
+  echo "starting installer jobs"
+  start_installer_job $WIN_JENKINS $WIN_JENKINS_USER $WIN_JENKINS_KEY $tag
+  start_installer_job $MAC_JENKINS $MAC_JENKINS_USER $MAC_JENKINS_KEY $tag
+fi
 
 popd > /dev/null
